Adapt extra/ and associated tests
