+++ src/extra/plugin/bind.js
@@ -1,6 +1,9 @@
 // This plugin replace Parsley default form behavior that auto bind its fields children
 // With this plugin you must register in constructor your form's fields and their constraints
 // You have this way a total javascript control over your form validation, and nothing needed in DOM
+
+import jQuery from 'jquery'; // Remove this line in ES3
+
 (function($){
 
 window.ParsleyConfig = $.extend(true, window.ParsleyConfig, { autoBind: false });
