+++ src/extra/validator/comparison.js
@@ -1,52 +1,43 @@
-(function () {
+// Load this after Parsley for additional comparison validators
+// Note: comparing with a reference isn't well supported and not recommended.
+import jQuery from 'jquery'; // Remove this line in ES3
+
+// gt, gte, lt, lte, notequalto extra validators
 var parseRequirement = function (requirement) {
   if (isNaN(+requirement))
-    return parseFloat($(requirement).val());
+    return parseFloat(jQuery(requirement).val());
   else
     return +requirement;
 };
 
-// gt, gte, lt, lte, notequalto extra validators
-window.ParsleyConfig = window.ParsleyConfig || {};
-window.ParsleyConfig.validators = window.ParsleyConfig.validators || {};
-
 // Greater than validator
-window.ParsleyConfig.validators.gt = {
-  fn: function (value, requirement) {
+window.Parsley.addValidator('gt', {
+  validateString: function (value, requirement) {
     return parseFloat(value) > parseRequirement(requirement);
   },
   priority: 32
-};
+});
 
 // Greater than or equal to validator
-window.ParsleyConfig.validators.gte = {
-  fn: function (value, requirement) {
+window.Parsley.addValidator('gte', {
+  validateString: function (value, requirement) {
     return parseFloat(value) >= parseRequirement(requirement);
   },
   priority: 32
-};
+});
 
 // Less than validator
-window.ParsleyConfig.validators.lt = {
-  fn: function (value, requirement) {
+window.Parsley.addValidator('lt', {
+  validateString: function (value, requirement) {
     return parseFloat(value) < parseRequirement(requirement);
   },
   priority: 32
-};
+});
 
 // Less than or equal to validator
-window.ParsleyConfig.validators.lte = {
-  fn: function (value, requirement) {
+window.Parsley.addValidator('lte', {
+  validateString: function (value, requirement) {
     return parseFloat(value) <= parseRequirement(requirement);
   },
   priority: 32
-};
-
-// Not equal to validator
-window.ParsleyConfig.validators.lte = {
-  fn: function (value, requirement) {
-    return parseFloat(value) <= parseRequirement(requirement);
-  },
-  priority: 32
-};
-})();
+});
