+++ src/extra/validator/dateiso.js
@@ -1,11 +1,8 @@
-// dateiso extra validator
-// Guillaume Potier
-window.ParsleyConfig = window.ParsleyConfig || {};
-window.ParsleyConfig.validators = window.ParsleyConfig.validators || {};
+// Load this after Parsley for additional comparison validators
 
-window.ParsleyConfig.validators.dateiso = {
-  fn: function (value) {
+window.Parsley.addValidator('dateiso', {
+  validateString: function (value) {
     return /^(\d{4})\D?(0[1-9]|1[0-2])\D?([12]\d|0[1-9]|3[01])$/.test(value);
   },
   priority: 256
-};
+});
